requirements.txt
aniso8601==6.0.0
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Extended==3.18.1
Flask-RESTful==0.3.7
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.1
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
Werkzeug==0.15.2
